+++ CHANGELOG.md
@@ -4,14 +4,23 @@ This project does its best to adhere to [Semantic Versioning](http://semver.org/
 
 
 --------
-###[0.1.1](N/A) - 2016-08-21
-####Changed
+###[0.1.2](N/A) - 2016-10-21
+#### Changed
+* Added unit test for non-interactive parameter parsing
+
+#### Fixed
+* Fixed an issue with parsing enums
+
+
+--------
+###[0.1.1](https://github.com/TeamworkGuy2/JParameter/commit/393f65924b375af3949767606e90ca1327e97376) - 2016-08-21
+#### Changed
 * Added jcollection-builders dependency
 * Updated jcollection-util dependency to latest 0.7.0 version
 
 
 --------
 ###[0.1.0](https://github.com/TeamworkGuy2/JParameter/commit/63df5669659db8195e2a6261d492fc2d234a3718) - 2016-06-26
-####Added
+#### Added
 * Versioning of existing code. Basic parameter parsing for command line arguments or interactive user input. Includes support for int, float, string, Path, and enum.
 * Updated to latest version of JCollectionFiller, (MapBuilder newMutableEnumNames() renamed mutableEnumNames())